requirements.txt
plotly==5.7.0
pyannote.core>=4.4
pyannote.database>=4.1.3